drivers/common: Extract dhcp package
Move DHCP lease types and functions from pkg/drivers/common into a
dedicated dhcp subpackage, preparing for deduplication of the setupIP
polling loop across vfkit, krunkit, qemu, and hyperkit drivers (#21093)

Pure code movement with renames to fit the new package context:
- DHCPEntry → dhcp.Entry
- GetIPAddressByMACAddress → dhcp.IPAddressForMAC
- parseDHCPdLeasesFile → parseLeases
- getIPAddressFromFile → ipAddressFromFile
